GPU Based Audio Transcriber/requirements.txt
PyQt5
openai-whisper
torch
pydub
ffmpeg-python
